Rename: user.entity entities폴더 안으로 이동, socket -> socket-room 변경
