+++ src/channels/entities/channel-invitation.entity.ts
@@ -1,4 +1,4 @@
-import { User } from './../../users/user.entity';
+import { User } from '../../users/entities/user.entity';
 import {
   Entity,
   PrimaryGeneratedColumn,
