+++ src/user-relation/dtos/create-user-relation.dto.ts
@@ -1,5 +1,5 @@
 import { UserRelationStatusEnum } from 'src/user-relation/enums/user-relation-status.enum';
-import { User } from 'src/users/user.entity';
+import { User } from 'src/users/entities/user.entity';
 
 export class CreateUserRelationDto {
   user: User;
