+++ src/user-relation/user-relation.controller.ts
@@ -3,7 +3,7 @@ import { UserRelationService } from './user-relation.service';
 import { ShowFriendRelationsDto } from './dtos/show-friend-relations.dto';
 import { AuthGuard } from 'src/auth/auth.guard';
 import { GetUser } from 'src/auth/user.decorator';
-import { User } from 'src/users/user.entity';
+import { User } from 'src/users/entities/user.entity';
 import { UserByIdPipe } from 'src/pipes/UserById.pipe';
 import { Serialize } from 'src/interceptors/serializer.interceptor';
 import { ShowUserIdDto } from './dtos/show-user-id.dto';
