+++ src/user-relation/user-relation.module.ts
@@ -3,7 +3,7 @@ import { UserRelationController } from './user-relation.controller';
 import { UserRelationService } from './user-relation.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserRelation } from './user-relation.entity';
-import { User } from 'src/users/user.entity';
+import { User } from 'src/users/entities/user.entity';
 import { UsersModule } from 'src/users/users.module';
 import { UserRelationGateway } from './user-relation.gateway';
